If mentor is set in query, hide panel and only load that one mentor
Show loading status on main video

Show message status on main video

Don't sort recommended questions to top in history
