+++ client/MentorPanel/src/components/video-panel.js
@@ -1,12 +1,13 @@
-import React from "react"
-import { useSelector, useDispatch } from 'react-redux';
-import { CircularProgress } from '@material-ui/core';
-import { Sms, SmsFailed, Star } from '@material-ui/icons'
+import React from 'react'
+import { useSelector, useDispatch } from 'react-redux'
+import { Star } from '@material-ui/icons'
 
 import { selectMentor } from 'src/redux/actions'
-import { STATUS_READY, STATUS_ERROR } from 'src/redux/store'
+import { STATUS_ERROR } from 'src/redux/store'
 
-import VideoThumbnail from "src/components/video-thumbnail";
+import VideoThumbnail from 'src/components/video-thumbnail'
+import LoadingSpinner from 'src/components/video-spinner'
+import MessageStatus from 'src/components/video-status'
 
 const VideoPanel = ({ isMobile }) => {
   const dispatch = useDispatch()
@@ -40,37 +41,10 @@ const VideoPanel = ({ isMobile }) => {
   )
 }
 
-const LoadingSpinner = ({ mentor }) => {
-  const question = useSelector(state => state.current_question)
-  if (question && question !== mentor.question) {
-    return <CircularProgress className='spinner' />
-  }
-  return <div></div>
-}
-
 const StarIcon = ({ mentor }) => {
   const faved_mentor = useSelector(state => state.faved_mentor)
   if (faved_mentor === mentor.id) {
     return <Star className='star-icon' fontSize='small' style={{ color: 'yellow' }} />
-  }
-  return <div></div>
-}
-
-const MessageStatus = ({ mentor }) => {
-  const next_mentor = useSelector(state => state.next_mentor)
-
-  if (mentor.is_off_topic) {
-    return <div></div>
-  }
-  if (mentor.status === STATUS_ERROR) {
-    return <SmsFailed className='message-notice' fontSize='small' style={{ color: 'red' }} />
-  }
-  if (mentor.status === STATUS_READY) {
-    const isNext = mentor.id === next_mentor
-    return <Sms
-      className={`message-notice ${isNext ? 'blink' : ''}`}
-      fontSize='small'
-      style={{ color: 'green' }} />
   }
   return <div></div>
 }
